+++ src/rest/blog.resource.ts
@@ -1,0 +1,10 @@
+import {BlogService} from '../service/blog.service';
+import {BaseResource} from './base/blog.resource';
+
+export class BlogResource extends BaseResource<BlogService> {
+
+    constructor() {
+        super('/blog', new BlogService());
+    }
+
+}
