+++ src/util/decorators.ts
@@ -1,7 +1,7 @@
 import 'reflect-metadata';
 import {PageController} from '../controller/page.controller';
 import {Express} from 'express';
-import {BaseResource} from '../rest/blog.rest';
+import {BaseResource} from '../rest/blog.resource';
 
 export function PageRoute(path: string) {
     return function (target: any, propertyKey: string, descriptor: PropertyDescriptor) {
